rename source to sink
